feat: add logout
