+++ apps/web/services/index.ts
@@ -1,5 +1,9 @@
+export * from './kakao';
+export * from './util';
+
 import { BE_URL } from '@/constants';
 import axios from 'axios';
+
 export const apiBe = axios.create({
   baseURL: `${BE_URL}/api`,
   timeout: 3_000,
